feat: #1 separate owner and customer into independent services
