--- src/main/kotlin/com/carava/carwash/BasicLoginApplication.kt
@@ -1,11 +1,0 @@
-package com.carava.carwash
-
-import org.springframework.boot.autoconfigure.SpringBootApplication
-import org.springframework.boot.runApplication
-
-@SpringBootApplication
-class BasicLoginApplication
-
-fun main(args: Array<String>) {
-	runApplication<BasicLoginApplication>(*args)
-}
