+++ src/main/kotlin/com/carava/carwash/CarwashCustomerApplication.kt
@@ -1,0 +1,11 @@
+package com.carava.carwash
+
+import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.boot.runApplication
+
+@SpringBootApplication
+class CarwashCustomerApplication
+
+fun main(args: Array<String>) {
+	runApplication<CarwashCustomerApplication>(*args)
+}
