+++ src/main/kotlin/com/carava/carwash/auth/dto/CheckUsernameResponseDto.kt
@@ -1,0 +1,5 @@
+package com.carava.carwash.auth.dto
+
+data class CheckUsernameResponseDto (
+    val isDuplicate: Boolean
+)
