+++ src/main/kotlin/com/carava/carwash/auth/dto/SignInRequestDto.kt
@@ -1,0 +1,6 @@
+package com.carava.carwash.auth.dto
+
+data class SignInRequestDto (
+    val email: String,
+    val password: String
+)
