+++ src/main/kotlin/com/carava/carwash/auth/dto/SignInResponseDto.kt
@@ -1,0 +1,7 @@
+package com.carava.carwash.auth.dto
+
+data class SignInResponseDto (
+    val accessToken: String,
+    val tokenType: String = "Bearer",
+    val expiresIn: Long
+)
