+++ src/main/kotlin/com/carava/carwash/auth/dto/SignUpRequestDto.kt
@@ -1,0 +1,6 @@
+package com.carava.carwash.auth.dto
+
+data class SignUpRequestDto (
+    val email: String,
+    val password: String
+)
