+++ src/main/kotlin/com/carava/carwash/auth/dto/SignUpResponseDto.kt
@@ -1,0 +1,6 @@
+package com.carava.carwash.auth.dto
+
+data class SignUpResponseDto (
+    val email: String,
+    val createdAt: String
+)
