--- src/main/kotlin/com/carava/carwash/customer/dto/CheckUsernameResponseDto.kt
@@ -1,5 +1,0 @@
-package com.carava.carwash.customer.dto
-
-data class CheckUsernameResponseDto (
-    val isDuplicate: Boolean
-)
