--- src/main/kotlin/com/carava/carwash/customer/dto/SignInRequestDto.kt
@@ -1,6 +1,0 @@
-package com.carava.carwash.customer.dto
-
-data class SignInRequestDto (
-    val email: String,
-    val password: String
-)
