--- src/main/kotlin/com/carava/carwash/customer/dto/SignInResponseDto.kt
@@ -1,7 +1,0 @@
-package com.carava.carwash.customer.dto
-
-data class SignInResponseDto (
-    val accessToken: String,
-    val tokenType: String = "Bearer",
-    val expiresIn: Long
-)
