--- src/main/kotlin/com/carava/carwash/customer/dto/SignUpRequestDto.kt
@@ -1,6 +1,0 @@
-package com.carava.carwash.customer.dto
-
-data class SignUpRequestDto (
-    val email: String,
-    val password: String
-)
