--- src/main/kotlin/com/carava/carwash/owner/dto/CheckUsernameResponseDto.kt
@@ -1,5 +1,0 @@
-package com.carava.carwash.owner.dto
-
-data class CheckUsernameResponseDto (
-    val isDuplicate: Boolean
-)
