--- src/main/kotlin/com/carava/carwash/owner/dto/SignInRequestDto.kt
@@ -1,6 +1,0 @@
-package com.carava.carwash.owner.dto
-
-data class SignInRequestDto (
-    val email: String,
-    val password: String
-)
