--- src/main/kotlin/com/carava/carwash/owner/dto/SignInResponseDto.kt
@@ -1,7 +1,0 @@
-package com.carava.carwash.owner.dto
-
-data class SignInResponseDto (
-    val accessToken: String,
-    val tokenType: String = "Bearer",
-    val expiresIn: Long
-)
