--- src/main/kotlin/com/carava/carwash/owner/dto/SignUpResponseDto.kt
@@ -1,6 +1,0 @@
-package com.carava.carwash.owner.dto
-
-data class SignUpResponseDto (
-    val email: String,
-    val createdAt: String
-)
