--- src/test/kotlin/com/carava/basicLogin/BasicLoginApplicationTests.kt
@@ -1,13 +1,0 @@
-package com.carava.basicLogin
-
-import org.junit.jupiter.api.Test
-import org.springframework.boot.test.context.SpringBootTest
-
-@SpringBootTest
-class BasicLoginApplicationTests {
-
-	@Test
-	fun contextLoads() {
-	}
-
-}
