+++ src/test/kotlin/com/carava/carwash/CarwashCustomerApplicationTests.kt
@@ -1,0 +1,13 @@
+package com.carava.carwash
+
+import org.junit.jupiter.api.Test
+import org.springframework.boot.test.context.SpringBootTest
+
+@SpringBootTest
+class CarwashCustomerApplicationTests {
+
+	@Test
+	fun contextLoads() {
+	}
+
+}
